# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AIOps Monitoring Middleware API
  version: 1.5.0
  description: Provider-neutral monitoring API plus authenticated site-agent snapshot and alert ingestion. Implementations must not expose LibreNMS payloads, tokens, database access, or renderer-specific options.
servers:
  - url: http://127.0.0.1:4310
paths:
  /api/v1/site-agent/batches:
    post:
      operationId: ingestSiteAgentBatch
      summary: Upload a snapshot using an agent identity bound by the server
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteAgentSnapshotBatch' }
      responses:
        '202': { description: New batch accepted }
        '200': { description: Duplicate batch accepted idempotently }
        '400': { description: Invalid or non-whitelisted batch fields }
        '401': { description: Missing or invalid agent credential }
        '403': { description: Payload identity conflicts with authenticated source }
        '409': { description: Batch or sequence conflict }
        '413': { description: Batch exceeds 2 MiB }
  /api/v1/site-agent/alert-events:
    post:
      operationId: ingestSiteAgentAlertEvents
      summary: Upload one normalized Alertmanager Webhook delivery
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteAgentAlertEventBatch' }
      responses:
        '202': { description: New alert delivery accepted }
        '200': { description: Duplicate delivery accepted idempotently }
        '400': { description: Invalid, unknown or sensitive alert fields }
        '401': { description: Missing or invalid agent credential }
        '403': { description: Payload identity conflicts with authenticated source }
        '409': { description: Delivery ID conflict }
        '413': { description: Request exceeds 1 MiB }
  /api/v1/site-agent/alert-snapshots:
    post:
      operationId: ingestSiteAgentAlertSnapshot
      summary: Upload the complete active Alertmanager alert set
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteAgentAlertSnapshotBatch' }
      responses:
        '202': { description: New snapshot accepted, including an older unapplied sequence }
        '200': { description: Duplicate snapshot accepted idempotently }
        '400': { description: Invalid, unknown or sensitive alert fields }
        '401': { description: Missing or invalid agent credential }
        '403': { description: Payload identity conflicts with authenticated source }
        '409': { description: Snapshot ID or sequence conflict }
        '413': { description: Request exceeds 1 MiB }
  /api/v1/cloud/monitoring/sources:
    get:
      operationId: listCloudMonitoringSources
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Sources restricted to the viewer credential tenant set }
        '401': { description: Missing or invalid platform viewer credential }
...
  /api/v1/monitoring/overview:
    get:
      operationId: getMonitoringOverview
      responses:
        '200':
          description: Normalized device and alert summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringOverview' }
        '502': { $ref: '#/components/responses/UpstreamError' }
...
  /api/v1/monitoring/alerts:
    get:
      operationId: listMonitoringAlerts
      parameters:
        - in: query
          name: state
          schema: { type: string, enum: [all, active, ack, unack, open], default: all }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Normalized alerts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringAlertList' }
  /api/v1/monitoring/alert-history:
    get:
      operationId: listMonitoringAlertHistory
      summary: Query a read-only LibreNMS alert log window of at most 31 days
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: to, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { in: query, name: deviceId, schema: { type: string, maxLength: 160 } }
      responses:
        '200':
          description: Paged normalized alert history without sensitive LibreNMS fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringAlertHistoryPage' }
        '400': { description: Invalid query or range longer than 31 days }
  /api/v1/monitoring/series/port-traffic:
    get:
      operationId: queryPortTrafficSeries
      parameters:
        - in: query
          name: deviceId
          required: true
          schema: { type: string, maxLength: 160 }
        - in: query
          name: portId
          required: true
          schema: { type: string, maxLength: 160 }
      responses:
        '200':
          description: Renderer-neutral rolling traffic series
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeSeriesFrame' }
        '404': { description: Port not found }
  /api/v1/monitoring/ports:
    get:
      operationId: listAllMonitoringPorts
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [all, up, down], default: all } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { in: query, name: sort, schema: { type: string, enum: [id, traffic, utilization, errors, discards], default: traffic } }
        - { in: query, name: order, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: Paginated normalized ports across all devices; use id/asc for stable full collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringPortPage' }
  /api/v1/monitoring/ports/rankings:
    get:
      operationId: rankMonitoringPorts
      parameters:
        - { in: query, name: metric, schema: { type: string, enum: [traffic, utilization, errors, discards], default: traffic } }
        - { in: query, name: status, schema: { type: string, enum: [all, up, down], default: all } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Ranked ports with traffic, utilization, errors and discards
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringPortRankingPage' }
        '400': { description: Invalid ranking metric }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    UpstreamError:
      description: Provider unavailable; upstream response and credentials are never returned
  schemas:
    ProviderDescriptor:
      type: object
      additionalProperties: false
      required: [id, name, configured]
      properties:
        id: { type: string }
        name: { type: string }
        configured: { type: boolean }
...
    MonitoringOverview:
      type: object
      additionalProperties: false
      required: [generatedAt, provider, devices, alerts]
      properties:
        generatedAt: { type: string, format: date-time }
        provider: { $ref: '#/components/schemas/ProviderDescriptor' }
        devices:
          type: object
          additionalProperties: false
          required: [total, up, down, states]
          properties:
            total: { type: integer, minimum: 0 }
            up: { type: integer, minimum: 0 }
            down: { type: integer, minimum: 0 }
            states:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
        alerts:
          type: object
          additionalProperties: false
          required: [total, active, severities]
          properties:
            total: { type: integer, minimum: 0 }
            active: { type: integer, minimum: 0 }
            severities:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
...
    MonitoringPort:
      type: object
      required: [id, deviceId, name, description, status, adminStatus, speedBps, rxBps, txBps]
      properties:
        id: { type: string }
        deviceId: { type: string }
        deviceName: { type: string }
        name: { type: string }
        description: { type: string }
        group: { type: string }
        status: { type: string }
        adminStatus: { type: string }
        macAddress: { type: string }
        speedBps: { type: [number, 'null'], minimum: 0 }
        rxBps: { type: [number, 'null'], minimum: 0 }
        txBps: { type: [number, 'null'], minimum: 0 }
        trafficBps: { type: number, minimum: 0 }
        utilizationPercent: { type: [number, 'null'], minimum: 0 }
        errors: { type: integer, minimum: 0 }
        discards: { type: integer, minimum: 0 }
        polledAt: { type: [string, 'null'], format: date-time }
        pollIntervalSeconds: { type: [number, 'null'], minimum: 0 }
...
    MonitoringAlert:
      type: object
      additionalProperties: false
      required: [id, deviceId, deviceName, ruleId, title, severity, state, acknowledged, openedAt, updatedAt, note]
      properties:
        id: { type: string }
        deviceId: { type: string }
        deviceName: { type: string }
        ruleId: { type: string }
        title: { type: string }
        severity: { type: string }
        state: { type: string }
        acknowledged: { type: boolean }
        openedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
        note: { type: string }
    MonitoringAlertList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringAlert' }
    MonitoringAlertHistory:
      type: object
      additionalProperties: false
      required: [id, deviceId, deviceName, ruleId, title, severity, state, occurredAt, description]
      properties:
        id: { type: string }
        deviceId: { type: string }
        deviceName: { type: string }
        ruleId: { type: string }
        title: { type: string, maxLength: 500 }
        severity: { type: string }
        state: { type: string }
        occurredAt: { type: [string, 'null'], format: date-time }
        description: { type: string, maxLength: 4000 }
    MonitoringAlertHistoryPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringAlertHistory' }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, minimum: 0 }
    MonitoringPortPage:
      type: object
      additionalProperties: false
      required: [generatedAt, total, page, pageSize, items]
      properties:
        generatedAt: { type: string, format: date-time }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringPort' }
    MonitoringRankedPort:
      allOf:
        - { $ref: '#/components/schemas/MonitoringPort' }
        - type: object
          required: [rank, rankingValue]
          properties:
            rank: { type: integer, minimum: 1 }
            rankingValue: { type: number }
    MonitoringPortRankingPage:
      type: object
      additionalProperties: false
      required: [generatedAt, metric, total, items]
      properties:
        generatedAt: { type: string, format: date-time }
        metric: { type: string, enum: [traffic, utilization, errors, discards] }
        total: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringRankedPort' }
    SiteAgentSnapshotBatch:
      type: object
      additionalProperties: false
      required: [schemaVersion, batchId, sequence, kind, observedAt, devices, ports]
      properties:
        schemaVersion: { type: string, const: '1.0' }
        batchId: { type: string, minLength: 1, maxLength: 160 }
        sequence: { type: integer, minimum: 1 }
        kind: { type: string, const: snapshot }
        observedAt: { type: string, format: date-time }
        sourceId: { type: string, description: Optional assertion; must equal authenticated source }
        tenantId: { type: string, description: Optional assertion; never used as authorization input }
        siteId: { type: string, description: Optional assertion; must equal authenticated site }
        devices:
          type: array
          maxItems: 1000
          items: { $ref: '#/components/schemas/SiteAgentDevice' }
        ports:
          type: array
          maxItems: 5000
          items: { $ref: '#/components/schemas/SiteAgentPort' }
    SiteAgentDevice:
      type: object
      additionalProperties: false
      required: [localDeviceId, name, status]
      properties:
        localDeviceId: { type: string, maxLength: 160 }
        name: { type: string, maxLength: 255 }
        hostname: { type: string, maxLength: 255 }
        ip: { type: string }
        status: { type: string, enum: [up, down, disabled, warning, unknown] }
        os: { type: string, maxLength: 120 }
        hardware: { type: string, maxLength: 255 }
        location: { type: string, maxLength: 255 }
        uptimeSeconds: { type: integer, minimum: 0 }
        polledAt: { type: string, format: date-time }
    SiteAgentPort:
      type: object
      additionalProperties: false
      required: [localPortId, localDeviceId, name, status]
      properties:
        localPortId: { type: string, maxLength: 160 }
        localDeviceId: { type: string, maxLength: 160 }
        name: { type: string, maxLength: 255 }
        description: { type: string, maxLength: 500 }
        group: { type: string, maxLength: 120 }
        status: { type: string, enum: [up, down, disabled, warning, unknown] }
        adminStatus: { type: string, enum: [up, down, disabled, warning, unknown] }
        macAddress: { type: string }
        speedBps: { type: number, minimum: 0 }
        rxBps: { type: number, minimum: 0 }
        txBps: { type: number, minimum: 0 }
        errors: { type: integer, minimum: 0 }
        discards: { type: integer, minimum: 0 }
        polledAt: { type: string, format: date-time }
    SiteAgentAlert:
      type: object
      additionalProperties: false
      required: [fingerprint, status, startsAt, labels, annotations]
      properties:
        fingerprint: { type: string, minLength: 1, maxLength: 160 }
        status: { type: string, enum: [firing, resolved, suppressed] }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        labels:
          type: object
          minProperties: 1
          maxProperties: 64
          additionalProperties: { type: string, maxLength: 512 }
        annotations:
          type: object
          maxProperties: 32
          additionalProperties: { type: string, maxLength: 2000 }
    SiteAgentAlertEventBatch:
      type: object
      additionalProperties: false
      required: [schemaVersion, deliveryId, kind, observedAt, status, alerts]
      properties:
        schemaVersion: { type: string, const: '1.0' }
        deliveryId: { type: string, minLength: 1, maxLength: 160 }
        kind: { type: string, const: alert-events }
        observedAt: { type: string, format: date-time }
        status: { type: string, enum: [firing, resolved] }
        groupKey: { type: string, maxLength: 512 }
        receiver: { type: string, maxLength: 160 }
        sourceId: { type: string, description: Optional assertion; must equal authenticated source }
        tenantId: { type: string, description: Optional assertion; never used as authorization input }
        siteId: { type: string, description: Optional assertion; must equal authenticated site }
        alerts:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/SiteAgentAlert' }
    SiteAgentAlertSnapshotBatch:
      type: object
      additionalProperties: false
      required: [schemaVersion, snapshotId, sequence, kind, observedAt, alerts]
      properties:
        schemaVersion: { type: string, const: '1.0' }
        snapshotId: { type: string, minLength: 1, maxLength: 160 }
        sequence: { type: integer, minimum: 1 }
        kind: { type: string, const: alert-snapshot }
        observedAt: { type: string, format: date-time }
        sourceId: { type: string, description: Optional assertion; must equal authenticated source }
        tenantId: { type: string, description: Optional assertion; never used as authorization input }
        siteId: { type: string, description: Optional assertion; must equal authenticated site }
        alerts:
          type: array
          maxItems: 5000
          items: { $ref: '#/components/schemas/SiteAgentAlert' }
    TimeSeriesFrame:
      type: object
      required: [id, title, unit, sampledAt, sampleMode, series]
      properties:
        id: { type: string }
        title: { type: string }
        unit: { type: string }
        sampledAt: { type: string, format: date-time }
        sampleMode: { type: string }
        pointMode: { type: string, enum: [bounded, all-available] }
        pointCount: { type: integer, minimum: 0 }
        series:
          type: array
          items:
            type: object
            required: [id, name, points]
            properties:
              id: { type: string }
              name: { type: string }
              points:
                type: array
                items:
                  type: array
                  prefixItems: [{ type: integer }, { type: [number, 'null'] }]
                  minItems: 2
                  maxItems: 2